Several problems addressed
1. Show contact`s gender in the profile of the contact.
2. Remove the "Send Gift" button in the detailed screen of an event,
since many people might be related to an event and then some combined
proposed presents search should be implemented.
3. Implement filtering of proposed presents by category - MUSIC, EVENT…
4. Implement filtering of proposed presents by price.
5. Show the price of the proposed gift and the store of the proposed
gift in the page with more information about a proposed gift.
6. Make a new initial page only with the GiftCase logo
7. Remove the extended sort menu on the pages which do not contain lists
8. Show the main page only when all data about the client is received
from FB and REST API log in
9. Show a button "Buy" in the page with more information about a
proposed gift. A screen in which the user can add mockup payment data is
shown. The screen should have button "Apply", the result of which will
be executing the requiest "Send gift" to the REST API.
10. Implement sorting and filtering in the proposed presents list.
11. Implement configuration menu from which a sort order and filtering
data can be defined.